Rename MasterOutputHooks to MixerHooks
